1、Fix "Scan Disk" user_uuid error that causes WebDAV unable to display scan list. 2、Fix the problem of abnormality when scanning non-existent folders.

+++ code/rest/matter_dao.go
@@ -518,9 +518,10 @@ func (this *MatterDao) Delete(matter *Matter) {
 		//delete from db.
 		db := core.CONTEXT.GetDB().Delete(&matter)
 		this.PanicError(db.Error)
-
-		//delete dir from disk.
-		util.DeleteEmptyDir(matter.AbsolutePath())
+		if util.PathExists(matter.AbsolutePath()) {
+			//delete dir from disk.
+			util.DeleteEmptyDir(matter.AbsolutePath())
+		}
 
 	} else {
